backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
supabase>=2.9.0
python-multipart>=0.0.16
pandas>=2.2.0
numpy>=2.0.0
ta>=0.10.2
pandas-ta>=0.3.14b0
aiohttp>=3.11.0
pydantic>=2.10.0
python-dotenv>=1.0.0
asyncpg>=0.30.0
sqlalchemy>=2.0.36
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4